+++ src/components/Contentful/StaticContent/Sidebar/presenter.js
@@ -6,13 +6,15 @@ import LibMarkdown from '../../../LibMarkdown'
 import PageLink from '../../PageLink'
 import Librarians from '../../../Librarians'
 import Related from '../../../Related'
+import ServicePoint from '../../../Contentful/ServicePoint'
+
 
 const Presenter = ({ cfStatic }) => (
-  <div key={`ContentfulSidebar_${cfStatic.sys.id}`} className='col-md-4 col-sm-5 col-xs-12 right' role='complementary'>
+  <div key={`ContentfulSidebar_${cfStatic.sys.id}`} className='col-md-4 col-sm-5 col-xs-12 right'>
     <PageLink className='button callout' cfPage={cfStatic.fields.callOutLink} />
     <LibMarkdown>{cfStatic.fields.shortDescription}</LibMarkdown>
     <Librarians netids={cfStatic.fields.contactPeople} />
-
+    <ServicePoint cfServicePoint={cfStatic.fields.servicePoint } />
     <Related className='p-pages' title='Related Pages' showImages={false}>{ cfStatic.fields.relatedPages }</Related>
   </div>
 )